requirements.txt
black==23.3.0
fastapi==0.98.0
pymongo==4.4.0
uvicorn==0.22.0
pydantic==1.10.9
locust==2.15.1
python-dateutil==2.8.2
